added profile page to root dir
